requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.26.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
gunicorn==21.2.0
openpyxl==3.1.2
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.6
google-generativeai==0.3.2
chromadb==0.4.22
lxml==5.1.0
scikit-learn==1.3.2
